Server implementation of frame control that supports random access, bi-directionals, and skip intervals for fast execution. Supplies the functionality needed to complete feature #438 "need a "scrub through" feature".
Summary of changes:

Added new commands to CompilerActor to do random frame access.
Added a FrameChange merger function that aggregates all the changes in a sequence of frames.
Added a ClientFrame generator to the FrameController that is "frame change" aware and should create correct ClientFrames regardless of direction or frame skip interval.
To remain temporarily compatible with the old compiler socket json, creates several "Legacy" classes to preserve the old format.
Adding a new field to frame that captures the changes to grid, robot location, and register changes and updating processor and related classes to create the new field.
Removed responseToJson from the SocketReponseConvertFlow because only the apply is ever used.  Pretty much wanted to do something different in the compiler flow and discovered this constraint was unneeded.
Refactored out the Frame.nopFrame method because it wasn't used anywhere.
Refactored Frame.index to remove the optional since its correctly created everywhere.
Some existing methods are depecated and will be removed with the scrub feature is completed.
Preserves the initial and final frames during skip.